discount 자료형 (int -> double)

+++ src/main/java/com/psh/mybook/model/book/BookEnrollParam.java
@@ -30,7 +30,7 @@ public class BookEnrollParam {
     @NotNull
     private int quantity;
 
-    private int discount;
+    private double discount;
 
     private Date regDate;
