+++ src/main/java/com/psh/mybook/model/book/BookUpdateParam.java
@@ -26,7 +26,7 @@ public class BookUpdateParam {
 
     private int quantity;
 
-    private int discount;
+    private double discount;
 
     private Date regDate;
